apimon: optimize plugin by adding no-retval dll hook modifier (#1757)
* if we don't need return status or output parameter values, we need install return hook for this hook entry
* for such entrues we may add no-retval modificator, i.e. 'name.dll,FuncName,no-retval,log,type1,type2'
* add some unit-tests for parse_entry function
