+++ src/libusermode/userhook.cpp
@@ -771,9 +771,9 @@ void userhook::request_usermode_hook(drakvuf_t drakvuf, const dll_view_t* dll, c
     dll_t* p_dll = reinterpret_cast<dll_t*>(const_cast<dll_view_t*>(dll));
 
     if (target->type == HOOK_BY_NAME)
-        p_dll->targets.emplace_back(target->function_name, target->clsid, callback, target->argument_printers, extra);
+        p_dll->targets.emplace_back(target->function_name, target->clsid, target->no_retval, callback, target->argument_printers, extra);
     else // HOOK_BY_OFFSET
-        p_dll->targets.emplace_back(target->function_name, target->clsid, target->offset, callback, target->argument_printers, extra);
+        p_dll->targets.emplace_back(target->function_name, target->clsid, target->offset, target->no_retval, callback, target->argument_printers, extra);
 }
 
 void userhook::register_plugin(drakvuf_t drakvuf, usermode_cb_registration reg)
@@ -926,166 +926,6 @@ bool drakvuf_stop_userhooks(drakvuf_t drakvuf)
 {
     return !userhook::is_supported(drakvuf) || userhook::get_instance(drakvuf).stop();
 }
-
-std::optional<HookActions> get_hook_actions(const std::string& str)
-{
-    if (str == "log")
-    {
-        return HookActions::empty().set_log();
-    }
-    else if (str == "log+stack")
-    {
-        return HookActions::empty().set_log().set_stack();
-    }
-
-    return std::nullopt;
-}
-
-namespace
-{
-std::optional<std::string> try_parse_token(std::stringstream& ss)
-{
-    const char SEPARATOR = ',';
-    std::string result;
-    if (!std::getline(ss, result, SEPARATOR) || result.empty())
-    {
-        return std::nullopt;
-    }
-    return result;
-}
-
-std::string parse_token(std::stringstream& ss)
-{
-    auto maybe_token = try_parse_token(ss);
-    if (!maybe_token)
-    {
-        throw std::runtime_error{"Expected a token"};
-    }
-    return *maybe_token;
-}
-
-std::unique_ptr<ArgumentPrinter> make_arg_printer(
-    const PrinterConfig& config,
-    const std::string& type,
-    const std::string& name)
-{
-    if (type == "lpstr" || type == "lpcstr" || type == "lpctstr")
-    {
-        return std::make_unique<AsciiPrinter>(name, config);
-    }
-    else if (type == "lpcwstr" || type == "lpwstr" || type == "bstr")
-    {
-        return std::make_unique<WideStringPrinter>(name, config);
-    }
-    else if (type == "punicode_string")
-    {
-        return std::make_unique<UnicodePrinter>(name, config);
-    }
-    else if (type == "pulong")
-    {
-        return std::make_unique<UlongPrinter>(name, config);
-    }
-    else if (type == "pulonglong")
-    {
-        return std::make_unique<UlongLongPrinter>(name, config);
-    }
-    else if (type == "lpvoid*")
-    {
-        return std::make_unique<PointerToPointerPrinter>(name, config);
-    }
-    else if (type == "refclsid" || type == "refiid")
-    {
-        return std::make_unique<GuidPrinter>(name, config);
-    }
-    else if (type == "binary16")
-    {
-        return std::make_unique<Binary16StringPrinter>(name, config);
-    }
-
-    return std::make_unique<ArgumentPrinter>(name, config);
-}
-
-
-std::vector<std::unique_ptr<ArgumentPrinter>> parse_arguments(
-        const PrinterConfig& config,
-        std::stringstream& ss)
-{
-    std::vector<std::unique_ptr<ArgumentPrinter>> argument_printers;
-
-    for (size_t arg_idx = 0; ; arg_idx++)
-    {
-        auto maybe_arg = try_parse_token(ss);
-        if (!maybe_arg) break;
-
-        const std::string arg = *maybe_arg;
-        std::string arg_name;
-        std::string arg_type;
-        const auto pos = arg.find_first_of(':');
-
-        if (pos == std::string::npos)
-        {
-            arg_name = std::string("Arg") + std::to_string(arg_idx);
-            arg_type = arg;
-        }
-        else
-        {
-            arg_name = arg.substr(0, pos);
-            arg_type = arg.substr(pos + 1);
-        }
-
-        argument_printers.emplace_back(make_arg_printer(config, arg_type, arg_name));
-    }
-    return argument_printers;
-}
-
-plugin_target_config_entry_t parse_entry(
-    std::stringstream& ss,
-    PrinterConfig& config)
-{
-    plugin_target_config_entry_t entry{};
-
-    entry.dll_name = parse_token(ss);
-    entry.function_name = parse_token(ss);
-    entry.type = HOOK_BY_NAME;
-
-    std::string log_strategy_or_offset;
-    std::string token = parse_token(ss);
-    if (token == "clsid")
-    {
-        entry.clsid = parse_token(ss);
-        log_strategy_or_offset = parse_token(ss);
-    }
-    else
-    {
-        log_strategy_or_offset = token;
-    }
-
-    std::optional<HookActions> actions = get_hook_actions(log_strategy_or_offset);
-    if (!actions)
-    {
-        entry.type = HOOK_BY_OFFSET;
-        try
-        {
-            entry.offset = std::stoull(log_strategy_or_offset, 0, 16);
-        }
-        catch (const std::logic_error& exc)
-        {
-            throw std::runtime_error{"Invalid offset"};
-        }
-
-        std::string strategy_name = parse_token(ss);
-        actions = get_hook_actions(strategy_name);
-        if (!actions)
-            throw std::runtime_error{"Invalid hook action"};
-    }
-
-    entry.actions = *actions;
-    entry.argument_printers = parse_arguments(config, ss);
-
-    return entry;
-}
-} // namespace
-
 
 void drakvuf_load_dll_hook_config(drakvuf_t drakvuf, const char* dll_hooks_list_path, bool print_no_addr, const hook_filter_t& hook_filter, wanted_hooks_t& wanted_hooks)
 {
