+++ src/libusermode/userhook.hpp
@@ -113,61 +113,10 @@
 
 #include <glib.h>
 #include "plugins/plugins_ex.h"
+#include "utils.hpp"
 #include "printers/printers.hpp"
 
 typedef event_response_t (*callback_t)(drakvuf_t drakvuf, drakvuf_trap_info* info);
-
-enum target_hook_type
-{
-    HOOK_BY_NAME,
-    HOOK_BY_OFFSET
-};
-
-struct HookActions
-{
-    bool log;
-    bool stack;
-
-    HookActions() : log{false}, stack{false} {}
-
-    static HookActions empty()
-    {
-        return HookActions{};
-    }
-    HookActions& set_log()
-    {
-        this->log = true;
-        return *this;
-    }
-    HookActions& set_stack()
-    {
-        this->stack = true;
-        return *this;
-    }
-};
-
-struct plugin_target_config_entry_t
-{
-    std::string dll_name;
-    target_hook_type type;
-    std::string function_name;
-    std::string clsid;
-    addr_t offset;
-    HookActions actions;
-    std::vector< std::unique_ptr< ArgumentPrinter > > argument_printers;
-
-    plugin_target_config_entry_t()
-        : dll_name(), type(), function_name(), offset(), actions(), argument_printers()
-    {}
-
-    plugin_target_config_entry_t(std::string&& dll_name, std::string&& function_name, addr_t offset, HookActions hook_actions, std::vector< std::unique_ptr< ArgumentPrinter > >&& argument_printers)
-        : dll_name(std::move(dll_name)), type(HOOK_BY_OFFSET), function_name(std::move(function_name)), offset(offset), actions(hook_actions), argument_printers(std::move(argument_printers))
-    {}
-
-    plugin_target_config_entry_t(std::string&& dll_name, std::string&& function_name, HookActions hook_actions, std::vector< std::unique_ptr< ArgumentPrinter > >&& argument_printers)
-        : dll_name(std::move(dll_name)), type(HOOK_BY_NAME), function_name(std::move(function_name)), offset(), actions(hook_actions), argument_printers(std::move(argument_printers))
-    {}
-};
 
 enum target_hook_state
 {
@@ -179,23 +128,50 @@ enum target_hook_state
 
 struct hook_target_entry_t
 {
-    vmi_pid_t pid;
+    vmi_pid_t pid = 0;
     target_hook_type type;
     std::string target_name;
     std::string clsid;
     addr_t offset;
+    bool no_retval{false};
     callback_t callback;
     const std::vector < std::unique_ptr < ArgumentPrinter > >& argument_printers;
     target_hook_state state;
-    drakvuf_trap_t* trap;
+    drakvuf_trap_t* trap = nullptr;
     void* plugin;
 
-    hook_target_entry_t(std::string target_name, std::string clsid, callback_t callback, const std::vector < std::unique_ptr < ArgumentPrinter > >& argument_printers, void* plugin)
-        : pid(0), type(HOOK_BY_NAME), target_name(target_name), clsid(clsid), offset(0), callback(callback), argument_printers(argument_printers), state(HOOK_FIRST_TRY), trap(nullptr), plugin(plugin)
+    hook_target_entry_t(std::string target_name,
+        std::string clsid,
+        bool no_retval,
+        callback_t callback,
+        const std::vector < std::unique_ptr < ArgumentPrinter > >& argument_printers,
+        void* plugin)
+        : type(HOOK_BY_NAME)
+        , target_name(std::move(target_name))
+        , clsid(std::move(clsid))
+        , offset(0)
+        , no_retval(no_retval)
+        , callback(callback)
+        , argument_printers(argument_printers)
+        , state(HOOK_FIRST_TRY)
+        , plugin(plugin)
     {}
 
-    hook_target_entry_t(std::string target_name, std::string clsid, addr_t offset, callback_t callback, const std::vector < std::unique_ptr < ArgumentPrinter > >& argument_printers, void* plugin)
-        : pid(0), type(HOOK_BY_OFFSET), target_name(target_name), clsid(clsid), offset(offset), callback(callback), argument_printers(argument_printers), state(HOOK_FIRST_TRY), trap(nullptr), plugin(plugin)
+    hook_target_entry_t(std::string target_name,
+        std::string clsid,
+        addr_t offset,
+        bool no_retval,
+        callback_t callback,
+        const std::vector < std::unique_ptr < ArgumentPrinter > >& argument_printers,
+        void* plugin)
+        : type(HOOK_BY_OFFSET)
+        , target_name(std::move(target_name))
+        , clsid(std::move(clsid))
+        , offset(offset)
+        , no_retval(no_retval)
+        , callback(callback)
+        , argument_printers(argument_printers)
+        , state(HOOK_FIRST_TRY), plugin(plugin)
     {}
 };
